made medieval survival work on android!

+++ src/include/filesystem/jsonReader.h
@@ -6,7 +6,7 @@ struct jsonContainer
 {
 	std::wstring value = std::wstring(L"");
 	std::vector<jsonContainer> children = std::vector<jsonContainer>();
-	int getChildIndex(const std::wstring& value) const;
+	size_t getChildIndex(const std::wstring& value) const;
 	jsonContainer getChild(const std::wstring& value) const;
 	jsonContainer operator[](const std::wstring& name) const;
 };
